refactor: myCart -> cart 로 변수명 수정
